Move actions handling to a separate class
Fixes #2287
