+++ packit/base_git.py
@@ -2,12 +2,11 @@
 # SPDX-License-Identifier: MIT
 
 import os
-import shlex
 from collections.abc import Iterable
 from importlib.metadata import version
 from logging import getLogger
 from pathlib import Path
-from typing import Callable, Optional
+from typing import Optional
 
 import git
 import requests
@@ -24,7 +23,7 @@ from specfile.exceptions import (
 from specfile.macro_definitions import CommentOutStyle
 from specfile.sections import Section
 
-from packit.actions import ActionName
+from packit.actions_handler import ActionsHandler
 from packit.command_handler import RUN_COMMAND_HANDLER_MAPPING, CommandHandler
 from packit.config import Config, RunCommandType
 from packit.config.common_package_config import MultiplePackages
@@ -61,6 +60,7 @@ class PackitRepositoryBase:
 
         self._handler_kls = None
         self._command_handler: Optional[CommandHandler] = None
+        self._actions_handler: Optional[ActionsHandler] = None
 
     @property
     def handler_kls(self):
@@ -77,6 +77,15 @@ class PackitRepositoryBase:
                 config=self.config,
             )
         return self._command_handler
+
+    @property
+    def actions_handler(self):
+        if not self._actions_handler:
+            self._actions_handler = ActionsHandler(
+                self.package_config,
+                self.command_handler,
+            )
+        return self._actions_handler
 
     @property
     def repository_cache(self) -> Optional[RepositoryCache]:
@@ -218,148 +227,6 @@ class PackitRepositoryBase:
         with commit_message_file(main_msg, msg, trailers) as commit_message:
             commit_args = ["-F", commit_message]
             self.local_project.git_repo.git.commit(*commit_args)
-
-    def run_action(
-        self,
-        actions: ActionName,
-        method: Optional[Callable] = None,
-        env: Optional[dict] = None,
-        *args,
-        **kwargs,
-    ):
-        """
-        Run the method in the self._with_action block.
-
-        Usage:
-
-        >   self._run_action(
-        >        action_name="sync", method=dg.sync_files, upstream_project=up.local_project
-        >   )
-        >   # If user provided custom command for the `sync`, it will be used.
-        >   # Otherwise, the method `dg.sync_files` will be used
-        >   # with parameter `upstream_project=up.local_project`
-        >
-        >   self._run_action(action_name="pre-sync")
-        >   # This will be used as an optional hook
-
-        :param actions: ActionName enum (Name of the action that can be overwritten
-                                                in the package_config.actions)
-        :param method: method to run if the action was not defined by user
-                    (if not specified, the action can be used for custom hooks)
-        :param args: args for the method
-        :param kwargs: kwargs for the method
-        """
-        if not method:
-            logger.debug(f"Running {actions} hook.")
-        if self.with_action(action=actions, env=env) and method:
-            method(*args, **kwargs)
-
-    def has_action(self, action: ActionName) -> bool:
-        """
-        Is the action defined in the config?
-        """
-        return action in self.package_config.actions
-
-    def get_commands_for_actions(self, action: ActionName) -> list[list[str]]:
-        """
-        Parse the following types of the structure and return list of commands in the form of list.
-
-        I)
-        action_name: "one cmd"
-
-        II)
-        action_name:
-          - "one cmd""
-
-        III)
-        action_name:
-          - ["one", "cmd"]
-
-
-        Returns [["one", "cmd"]] for all of them.
-
-        :param action: str or list[str] or list[list[str]]
-        :return: list[list[str]]
-        """
-        configured_action = self.package_config.actions[action]
-        if isinstance(configured_action, str):
-            configured_action = [configured_action]
-
-        if not isinstance(configured_action, list):
-            raise ValueError(
-                f"Expecting 'str' or 'list' as a command, got '{type(configured_action)}'. "
-                f"The value: {configured_action}",
-            )
-
-        parsed_commands = []
-        for cmd in configured_action:
-            if isinstance(cmd, str):
-                parsed_commands.append(shlex.split(cmd))
-            elif isinstance(cmd, list):
-                parsed_commands.append(cmd)
-            else:
-                raise ValueError(
-                    f"Expecting 'str' or 'list' as a command, got '{type(cmd)}'. "
-                    f"The value: {cmd}",
-                )
-        return parsed_commands
-
-    def with_action(self, action: ActionName, env: Optional[dict] = None) -> bool:
-        """
-        If the action is defined in the self.package_config.actions,
-        we run it and return False (so we can skip the if block)
-
-        If the action is not defined, return True.
-
-        Usage:
-
-        >   if self._with_action(action_name="patch"):
-        >       # Run default implementation
-        >
-        >   # Custom command was run if defined in the config
-
-        Context manager is currently not possible without ugly hacks:
-        https://stackoverflow.com/questions/12594148/skipping-execution-of-with-block
-        https://www.python.org/dev/peps/pep-0377/ (rejected)
-
-        :param action: ActionName enum (Name of the action that can be overwritten
-                                                in the package_config.actions)
-        :param env: dict with env vars to set for the command
-        :return: True, if the action is not overwritten, False when custom command was run
-        """
-        logger.debug(f"Running {action}.")
-        if action in self.package_config.actions:
-            commands_to_run = self.get_commands_for_actions(action)
-            logger.info(f"Using user-defined script for {action}: {commands_to_run}")
-            for cmd in commands_to_run:
-                self.command_handler.run_command(command=cmd, env=env, print_live=True)
-            return False
-        logger.debug(f"Running default implementation for {action}.")
-        return True
-
-    def get_output_from_action(
-        self,
-        action: ActionName,
-        env: Optional[dict] = None,
-    ) -> Optional[list[str]]:
-        """
-        Run self.actions[action] command(s) and return their outputs.
-        """
-        if action not in self.package_config.actions:
-            return None
-
-        commands_to_run = self.get_commands_for_actions(action)
-
-        logger.info(f"Using user-defined script for {action}: {commands_to_run}")
-        return [
-            self.command_handler.run_command(
-                cmd,
-                return_output=True,
-                env=env,
-                print_live=True,
-            ).stdout
-            for cmd in commands_to_run
-        ]
 
     def specfile_add_patches(
         self,
