+++ tests/integration/test_changelog_helper.py
@@ -203,10 +203,14 @@ def test_update_distgit_changelog_entry_action_pass_env_vars(
         "PACKIT_PROJECT_UPSTREAM_TAG": "0.1.0",
         "PACKIT_PROJECT_PREVIOUS_VERSION": "0.0.0",
     }
-    flexmock(upstream).should_receive("get_output_from_action").with_args(
+    flexmock(upstream.actions_handler).should_receive(
+        "get_output_from_action",
+    ).with_args(
         ActionName.changelog_entry,
         env=expected_env,
-    ).and_return("- entry").once()
+    ).and_return(
+        "- entry",
+    ).once()
 
     ChangelogHelper(upstream, downstream, package_config).update_dist_git(
         upstream_tag="0.1.0",
